Modified contacts deletion

+++ controllers/contactController.js
@@ -5,7 +5,7 @@ const Contact = require("../models/contactModel");
 //@route GET /api/contacts
 //@access private 
 const getContacts = asyncHandler(async (req,res)=>{
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id:req.user.id});
     res.status(200).json(contacts);
 });
 
@@ -22,7 +22,10 @@ const createContact = asyncHandler(async (req,res)=>{
         throw new Error('All fields are required');
     }
     const contact = await Contact.create({
-        name,email,phone
+        name,
+        email,
+        phone,
+        user_id: req.user.id
     });
     res.status(201).json(contact);
 });
@@ -35,7 +38,7 @@ const getContact = asyncHandler(async (req,res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
-        throw Error("Contact not found");
+        throw new Error("Contact not found");
     }
     res.status(200).json(contact);
 });
@@ -48,11 +51,17 @@ const updateContact = asyncHandler(async (req,res)=>{
    const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
-        throw Error("Contact not found");
+        throw new Error("Contact not found");
+    }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);//forbidden
+        throw new Error("Action not permitted");
     }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
-        req.body,{new: true}
+        req.body,
+        {new: true}
     );
     res.status(200).json(updatedContact);
 });
@@ -65,7 +74,11 @@ const deleteContact = asyncHandler(async (req,res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
-        throw Error("contact not found");
+        throw new Error("contact not found");
+    }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403); //forbidden
+        throw new Error("Action not permitted");
     }
     await Contact.findByIdAndDelete(contact);
     res.status(200).json({message:`Deleted contact of id ${req.params.id}`});
